Refactor to separate App from Index
